pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "apexcoach"
version = "0.1.0"
description = "Rule-based coaching assistant MVP for Apex"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
  "numpy>=1.24",
  "opencv-python>=4.8",
  "Pillow>=10.0",
  "PyYAML>=6.0",
  "mss>=9.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
apexcoach = "apexcoach.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q -p no:cacheprovider"
